+++ Aircraft Simulation/Aircraft Simulation/time.hpp
@@ -21,12 +21,12 @@ public:
     
     // Getters
     systemTime getSystemTime(void) { return systemtime; }
-    float getSimTime(void)         { return curTime; }
+    double getSimTime(void)         { return curTime; }
     bool  performPrint(void)       { return print; }
     bool  performSave(void)        { return save; }
     bool  performDynamics(void)    { return dynamics; }
     int   getCount(void)           { return counter; } // get counter at clock_dt rate
-    float getDeltaTime(void)       { return dt; }      // get delta time of simulation
+    double getDeltaTime(void)       { return dt; }      // get delta time of simulation
     
     // Functions
     virtual bool update(void);
@@ -46,11 +46,11 @@ private:
     double timeSinceLastSave;
     double timeSinceLastClockDt;
     
-    float curTime;
-    float runTime;
-    float dynamicsInterval;
-    float printInterval;
-    float saveInterval;
+    double curTime;
+    double runTime;
+    double dynamicsInterval;
+    double printInterval;
+    double saveInterval;
     
     // Getters
     bool print;
